fix: transferred code, added comments
